Added sprites and blocks

+++ Minigin/Texture2D.cpp
@@ -6,7 +6,7 @@ dae::Texture2D::~Texture2D()
 	SDL_DestroyTexture(m_texture);
 }
 
-glm::ivec2 dae::Texture2D::GetSize() const
+glm::vec2 dae::Texture2D::GetSize() const
 {
 	SDL_Rect dst;
 	SDL_QueryTexture(GetSDLTexture(), nullptr, nullptr, &dst.w, &dst.h);
