+++ Minigin/Texture2D.h
@@ -14,7 +14,7 @@ namespace dae
 		explicit Texture2D(SDL_Texture* texture);
 		~Texture2D();
 
-		glm::ivec2 GetSize() const;
+		glm::vec2 GetSize() const;
 
 		Texture2D(const Texture2D &) = delete;
 		Texture2D(Texture2D &&) = delete;
